Add hack to handle the two different response status when reading.
When the `accept` header is present, it should return a 200 response with data, including the permission, but it returns a 204 as the documentations says when the header is not present.
I guess it's a bug on their side and it eventually will be fixed, and we can get the permission from the model.